requirements.txt
absl-py==2.3.0
action-msgs==1.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiosignal==1.3.2
ament-cmake-test==1.3.10
ament-index-python==1.4.0
ament-package==0.14.0
angles==1.15.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
arm_pytorch_utilities==0.4.3
asttokens==3.0.0
attrs==25.3.0
av==14.4.0
beautifulsoup4==4.13.4
blinker==1.9.0
bond==3.0.2
builtin-interfaces==1.2.1
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
cmake==4.0.3
composition-interfaces==1.2.1
contourpy==1.3.2
crc32c==2.7.1
cycler==0.12.1
dacite==1.9.2
datasets==3.6.0
decorator==5.2.1
deepdiff==8.5.0
diagnostic-msgs==4.2.4
diagnostic-updater==4.0.0
diffusers==0.33.1
dill==0.3.8
dm-env==1.6
dm-tree==0.1.9
dm_control==1.0.31
docstring_parser==0.16
domain-coordinator==0.10.0
donfig==0.8.1.post1
dotenv==0.9.9
draccus==0.10.0
dwb-msgs==1.1.16
einops==0.8.1
etils==1.12.2
evdev==1.9.2
executing==2.2.0
Farama-Notifications==0.0.4
fast_kinematics==0.2.2
feetech-servo-sdk==1.0.0
filelock==3.18.0
Flask==3.1.1
fonttools==4.58.4
frozenlist==1.7.0
fsspec==2025.3.0
gdown==5.2.0
geometry-msgs==4.2.4
gitdb==4.0.12
GitPython==3.1.44
glfw==2.9.0
google-auth==2.40.3
google-genai==1.20.0
grpcio==1.73.0
gymnasium==0.29.1
h11==0.16.0
h5py==3.14.0
hf-xet==1.1.3
hf_transfer==0.1.9
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
inquirerpy==0.3.4
interactive-markers==2.3.2
ipython==9.3.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joint-state-publisher==2.4.0
jsonlines==4.0.0
kiwisolver==1.4.8
labmaze==1.0.6
laser-geometry==2.4.0
launch==1.0.6
launch-ros==0.19.7
launch-testing==1.0.6
launch-xml==1.0.6
launch-yaml==1.0.6
-e git+https://github.com/huggingface/lerobot.git@8fc18be0659eace8a237896a523526b7b1d7886c#egg=lerobot
-e git+ssh://git@github.com/pravsels/lerobot-sim2real.git@cdf3ee6d09f8ae8e690aaff51b8d1d1420a86b52#egg=lerobot_sim2real
lifecycle-msgs==1.2.1
llvmlite==0.44.0
lxml==5.4.0
mani-skill-nightly==2025.6.13.2242
map-msgs==2.1.0
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mergedeep==1.3.4
message-filters==4.3.5
mplib==0.1.1
mpmath==1.3.0
mujoco==3.3.3
multidict==6.4.4
multiprocess==0.70.16
mypy_extensions==1.1.0
nav-2d-msgs==1.1.16
nav-msgs==4.2.4
nav2-common==1.1.16
nav2-msgs==1.1.16
networkx==3.5
numba==0.61.2
numcodecs==0.16.1
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1749430489133/work/dist/numpy-2.3.0-cp311-cp311-linux_x86_64.whl#sha256=b312303ab84dd598256572796538df999b53da5aab9918f0490de0e792707309
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-ml-py==12.575.51
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
omegaconf==2.3.0
opencv-python==4.10.0
opencv-python-headless==4.10.0
orderly-set==5.4.1
osrf-pycommon==2.0.2
packaging @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_packaging_1745345660/work
pandas==2.3.0
parso==0.8.4
pexpect==4.9.0
pfzy==0.3.4
pillow==11.2.1
platformdirs==4.3.8
ply @ file:///home/conda/feedstock_root/build_artifacts/ply_1733239724146/work
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.6
pydantic_core==2.33.2
Pygments==2.19.1
pymunk==6.11.1
pynput==1.8.1
pynvml==12.0.0
PyOpenGL==3.1.9
pyparsing==3.2.3
pyperclip==1.9.0
PyQt5==5.15.9
PyQt5-sip==12.12.2
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-xlib==0.33
pytorch-kinematics==0.7.5
pytorch-seed==0.2.0
pytz==2025.2
PyYAML==6.0.2
pyyaml-include==1.4.1
pyzmq==27.0.0
rcl-interfaces==1.2.1
rclpy==3.3.14
rcutils==5.1.6
regex==2024.11.6
requests==2.32.4
rerun-sdk==0.23.3
resource-retriever==3.1.2
rich==14.0.0
rmw-dds-common==1.6.0
rosgraph-msgs==1.2.1
rosidl-adapter==3.1.5
rosidl-cli==3.1.5
rosidl-cmake==3.1.5
rosidl-generator-c==3.1.5
rosidl-generator-cpp==3.1.5
rosidl-generator-py==0.14.4
rosidl-parser==3.1.5
rosidl-typesupport-c==2.0.1
rosidl-typesupport-cpp==2.0.1
rosidl-typesupport-fastrtps-c==2.2.2
rosidl-typesupport-fastrtps-cpp==2.2.2
rosidl-typesupport-introspection-c==3.1.5
rosidl-typesupport-introspection-cpp==3.1.5
rpyutils==0.2.1
rsa==4.9.1
safetensors==0.5.3
sapien==3.0.0b1
scipy==1.15.3
sensor-msgs==4.2.4
sensor-msgs-py==4.2.4
sentry-sdk==2.30.0
setproctitle==1.3.6
shtab==1.7.2
sip @ file:///home/conda/feedstock_root/build_artifacts/sip_1697300428897/work
six==1.17.0
smclib==3.0.2
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
stack-data==0.6.3
statistics-msgs==1.2.1
std-msgs==4.2.4
std-srvs==4.2.4
sympy==1.14.0
tabulate==0.9.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==3.1.0
tf2-geometry-msgs==0.25.7
tf2-msgs==0.25.7
tf2-py==0.25.7
tf2-ros-py==0.25.7
toml @ file:///home/conda/feedstock_root/build_artifacts/toml_1734091811753/work
tomli @ file:///home/conda/feedstock_root/build_artifacts/tomli_1733256695513/work
toppra==0.6.3
torch==2.7.1
torchcodec==0.4.0
torchvision==0.22.1
tqdm==4.67.1
traitlets==5.14.3
transforms3d==0.4.2
trimesh==4.6.12
triton==3.3.1
typeguard==4.4.3
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tyro==0.9.24
tzdata==2025.2
unique-identifier-msgs==2.2.1
urllib3==2.4.0
visualization-msgs==4.2.4
wandb==0.20.1
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.1
zarr==3.0.8
zipp==3.23.0
